# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinanceOS Dashboard API
servers:
  - url: /api
    description: Base API path
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: |
        Session cookie set by POST /api/auth/login. Required for all
        protected routes (everything except /api/healthz, /api/health,
        /api/auth/*, and /api/pipeline/* which uses its own auth).
    pipelineRefreshToken:
      type: apiKey
      in: header
      name: x-refresh-token
      description: |
        Shared webhook token for POST /api/pipeline/refresh when called by
        the external pipeline job. Alternatively, an authenticated session
        with the `pipeline_refresh` permission is accepted.

  # ── Parameters ──────────────────────────────────────────────────────────────
...
  schemas:
...
    ApiEnvelope:
      type: object
      description: Standard JSON envelope returned by all successful responses
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          description: Response payload (shape varies by endpoint)
        source:
          type: string
          enum: [live, mock, cache, db]
          description: Where the data was sourced from
        ts:
          type: string
          format: date-time
          description: Server timestamp (ISO 8601)
      required: [ok, data, ts]
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          description: Machine-readable error code (optional)
        ts:
          type: string
          format: date-time
      required: [ok, error, ts]

    # Auth
...
  responses:
    Unauthorized:
      description: Not authenticated — session cookie missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /drafts:
    post:
      operationId: createDraft
      tags: [ReportDrafts]
      summary: Create a new report draft
      description: Builds a live report, generates analysis, creates a draft, and persists commentary. Requires editor role (admin, cfo, controller).
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [template, period, entities]
              properties:
                template:
                  type: string
                period:
                  type: string
                entities:
                  oneOf:
                    - type: string
                      enum: [all]
                    - type: array
                      items:
                        type: string
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Invalid input or unauthorized role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listDrafts
      tags: [ReportDrafts]
      summary: List drafts for a template and period
      security:
        - sessionCookie: []
      parameters:
        - name: template
          in: query
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of drafts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Missing required query params
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /drafts/{draftId}:
    get:
      operationId: getDraft
      tags: [ReportDrafts]
      summary: Get a single draft by ID
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /drafts/{draftId}/edits:
    patch:
      operationId: saveDraftEdits
      tags: [ReportDrafts]
      summary: Save narrative edits to a draft
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [editableContent]
              properties:
                editableContent:
                  type: object
                changeSummary:
                  type: string
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Invalid input or approved/superseded draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drafts/{draftId}/versions:
    get:
      operationId: getDraftVersions
      tags: [ReportDrafts]
      summary: Get version history for a draft
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /drafts/{draftId}/restore:
    post:
      operationId: restoreDraftVersion
      tags: [ReportDrafts]
      summary: Restore a prior version
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [versionNumber]
              properties:
                versionNumber:
                  type: integer
      responses:
        "200":
          description: Draft after restore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /drafts/{draftId}/submit:
    post:
      operationId: submitDraftForReview
      tags: [ReportDrafts]
      summary: Submit draft for review (draft → ready_for_review)
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /drafts/{draftId}/approve:
    post:
      operationId: approveDraft
      tags: [ReportDrafts]
      summary: Approve a draft (admin or cfo only)
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Approved draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Stale draft or invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Insufficient role (requires admin or cfo)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drafts/{draftId}/preview:
    get:
      operationId: getDraftPreview
      tags: [ReportDrafts]
      summary: Render draft HTML preview with narrative context
      security:
        - sessionCookie: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Preview data including HTML
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Commentary ───────────────────────────────────────────────────────────────
  /drafts/commentary:
    get:
      operationId: getCommentary
      tags: [ReportDrafts]
      summary: Get commentary entries for an entity, period, and template
      security:
        - sessionCookie: []
      parameters:
        - name: entity
          in: query
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: true
          schema:
            type: string
        - name: template
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Commentary entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: saveCommentary
      tags: [ReportDrafts]
      summary: Create or update a commentary entry (management_commentary or recommended_action only)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entitySlug, reportingPeriod, templateId, sectionKey, commentaryType, content]
              properties:
                entitySlug:
                  type: string
                reportingPeriod:
                  type: string
                templateId:
                  type: string
                sectionKey:
                  type: string
                commentaryType:
                  type: string
                  enum: [management_commentary, recommended_action]
                content:
                  type: string
                sortOrder:
                  type: integer
                existingId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Saved commentary entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Invalid input or forbidden type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /drafts/commentary/reorder:
    post:
      operationId: reorderCommentary
      tags: [ReportDrafts]
      summary: Reorder commentary entries
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /drafts/commentary/{commentaryId}/toggle:
    patch:
      operationId: toggleCommentary
      tags: [ReportDrafts]
      summary: Toggle whether a commentary block is included in the report
      security:
        - sessionCookie: []
      parameters:
        - name: commentaryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [included]
              properties:
                included:
                  type: boolean
      responses:
        "200":
          description: Updated commentary entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /drafts/commentary/{commentaryId}:
    delete:
      operationId: deleteCommentary
      tags: [ReportDrafts]
      summary: Delete a commentary entry (management_commentary and recommended_action only)
      security:
        - sessionCookie: []
      parameters:
        - name: commentaryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "400":
          description: Cannot delete FinanceOS analysis entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /drafts/commentary/{commentaryId}/approve:
    post:
      operationId: approveCommentary
      tags: [ReportDrafts]
      summary: Approve a commentary entry (admin or cfo only)
      security:
        - sessionCookie: []
      parameters:
        - name: commentaryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Approved commentary entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"